requirements.txt
python-dotenv
discord.py
certifi
tk